Update package version to 4.4.4: Core Update

+++ Runtime/Graphics/Scripts/SCR_LightAnimation.cs
@@ -1,6 +1,6 @@
 namespace Core.Graphics
 {
-    public enum LightAnimation : byte
+    public enum LightAnimation : int
     {
         DEFAULT,
         FLICKER_GENERIC,
